+++ django-frontend/src/frontend/static/js/edit_profile.js
@@ -1,0 +1,27 @@
+const updateUsername = async (newUsername) => {
+    const url = `http://localhost:3000/api/users/${userId}/`;
+    const data = {
+        username: newUsername,
+    };
+    
+    try {
+        const response = await fetch(url, {
+            method: 'PUT',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify(data),
+        });
+
+        if (response.ok) {
+            const updatedUser = await response.json();
+            console.log('User updated successfully:', updatedUser);
+        } else {
+            console.error('Failed to update user:', response.statusText);
+        }
+    } catch (error) {
+        console.error('Error:', error);
+    }
+};
+
+export { updateUsername };
