+++ django-frontend/src/frontend/templates/profile.html
@@ -1,5 +1,5 @@
 <script>
-	var userId = {{user.id}};
+	var userId = {{ user.id }};
 </script>
 <div class="container justify-content-center">
 	<div class="header d-flex justify-content-center align-items-center">
