+++ django-frontend/src/frontend/templates/topbar.html
@@ -6,7 +6,7 @@
 	</div>
 	<!-- account full name -->
 	<div class="account-user d-flex align-items-center">
-		<div class="account-username">{{ user.username }}</div>
+		<div class="account-username" id="accountUsername">{{ user.username }}</div>
 	</div>
 	<!-- logout button -->
 	<div class="account-logout">
