+++ drf-api-gateway/src/user/views.py
@@ -1,7 +1,6 @@
 from django.http import HttpResponse
 import requests
 from django.views import View
-from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
 from django.utils.decorators import method_decorator
